# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Malamh — Facial Consent Registry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /internal/faces:
    get:
      operationId: listFaces
      tags: [faces]
      summary: List current user's registered faces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Face"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFace
      tags: [faces]
      summary: Register a new face
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Face"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internal/faces/{id}:
    get:
      operationId: getFace
      tags: [faces]
      summary: Get a face by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Face"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFace
      tags: [faces]
      summary: Update face consent level or label
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFaceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Face"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFace
      tags: [faces]
      summary: Delete a face (and remove from AWS if indexed)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internal/faces/{id}/activity:
    get:
      operationId: getFaceActivity
      tags: [faces]
      summary: Get activity logs for a specific face
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessLog"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internal/embed:
    post:
      operationId: embedImage
      tags: [faces]
      summary: Generate embedding from a single image (multipart)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbedImageBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /keys:
    get:
      operationId: listApiKeys
      tags: [keys]
      summary: List user's API keys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createApiKey
      tags: [keys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [keys]
      summary: Delete an API key
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Consent Tokens ───────────────────────────────────────────────────────────
...
  /monitor/scan:
    post:
      operationId: runMonitorScan
      tags: [monitor]
      summary: Run a web scan for face appearances (MONITOR+ plan)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "403":
          description: Plan upgrade required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /monitor/results:
    get:
      operationId: listScanResults
      tags: [monitor]
      summary: List scan results for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /monitor/results/{id}:
    patch:
      operationId: updateScanResult
      tags: [monitor]
      summary: Update scan result status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScanResultBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Public API (for AI companies) ───────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ─── Auth ────────────────────────────────────────────────────────────────────
...
    Face:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        consentLevel:
          type: string
          enum: [BLOCKED, TOKEN_REQUIRED, OPEN]
        label:
          type: string
          nullable: true
        verified:
          type: boolean
        awsFaceId:
          type: string
          nullable: true
        embedding:
          type: string
          nullable: true
          description: JSON-stringified 512-d float array (only returned to owner)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, consentLevel, verified, createdAt, updatedAt]
    CreateFaceBody:
      type: object
      properties:
        embedding:
          type: string
          description: JSON-stringified 512-d float array
        consentLevel:
          type: string
          enum: [BLOCKED, TOKEN_REQUIRED, OPEN]
        label:
          type: string
          nullable: true
        verified:
          type: boolean
      required: [embedding, consentLevel]
    UpdateFaceBody:
      type: object
      properties:
        consentLevel:
          type: string
          enum: [BLOCKED, TOKEN_REQUIRED, OPEN]
        label:
          type: string
          nullable: true
    EmbedImageBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data
        mimeType:
          type: string
          description: Image MIME type e.g. image/jpeg
      required: [imageBase64]
    EmbedResponse:
      type: object
      properties:
        embedding:
          type: string
          description: JSON-stringified 512-d float array
        mock:
          type: boolean
      required: [embedding]
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: Full key only shown once on creation, otherwise masked
        active:
          type: boolean
        usageCount:
          type: number
        lastUsed:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, key, active, usageCount, createdAt]
    CreateApiKeyBody:
      type: object
      properties:
        name:
          type: string
      required: [name]

    # ─── Consent Tokens ───────────────────────────────────────────────────────────
...
    AccessLog:
      type: object
      properties:
        id:
          type: string
        faceId:
          type: string
          nullable: true
        requesterName:
          type: string
        requesterIp:
          type: string
        action:
          type: string
          enum: [checked, blocked, allowed, token_issued, token_used, no_match]
        matchScore:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, requesterName, requesterIp, action, matchScore, createdAt]
...
    ScanResult:
      type: object
      properties:
        id:
          type: string
        sourceUrl:
          type: string
        sourceDomain:
          type: string
        pageTitle:
          type: string
          nullable: true
        matchScore:
          type: number
        screenshotUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [NEW, SEEN, TAKEDOWN_SENT, RESOLVED, IGNORED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sourceUrl, sourceDomain, matchScore, status, createdAt, updatedAt]
    ScanResponse:
      type: object
      properties:
        success:
          type: boolean
        newResults:
          type: number
        results:
          type: array
          items:
            $ref: "#/components/schemas/ScanResult"
      required: [success, newResults, results]
    UpdateScanResultBody:
      type: object
      properties:
        status:
          type: string
          enum: [NEW, SEEN, TAKEDOWN_SENT, RESOLVED, IGNORED]
      required: [status]

    # ─── Public API ───────────────────────────────────────────────────────────────
...